feat(medium): Repair PR #9162: URL handling improvements and architectural fixes (#9253)
