+++ lib/env.ts
@@ -22,8 +22,14 @@ const envSchema = z
       }, z.boolean())
       .default(false),
     CI: z.string().optional(),
-    GOOGLE_DOC_WORKOUT_URL: z.string().url().optional().or(z.literal('')),
-    GOOGLE_DOC_IFRAME_URL: z.string().url().optional().or(z.literal('')),
+    GOOGLE_DOC_WORKOUT_URL: z.preprocess(
+      (str) => (str === '' ? undefined : str),
+      z.string().url().optional()
+    ),
+    GOOGLE_DOC_IFRAME_URL: z.preprocess(
+      (str) => (str === '' ? undefined : str),
+      z.string().url().optional()
+    ),
     NEXT_PUBLIC_USE_NATIVE_TABLE: z
       .preprocess((val) => {
         if (typeof val === 'string') return val.toLowerCase() === 'true'
